+++ tests/layout/test_checkbox.py
@@ -1,7 +1,7 @@
 """
 Tests to verify a single (boolean) checkbox is rendered correctly.
-
 """
+
 from tbxforms.layout import (
     Field,
     Layout,
@@ -26,9 +26,7 @@ def test_validation_error_attributes(snapshot_html):
 def test_checkbox_size(snapshot_html):
     """Verify size of the checkbox can be changed from the default."""
     form = CheckboxForm()
-    form.helper.layout = Layout(
-        Field("accept", context={"checkboxes_small": True})
-    )
+    form.helper.layout = Layout(Field.checkbox("accept", small=True))
     assert render_form(form) == snapshot_html