+++ tests/layout/test_fieldset.py
@@ -1,11 +1,7 @@
 """
 Tests to verify fieldsets are rendered correctly.
 """
-from tbxforms.layout import (
-    Fieldset,
-    Layout,
-    Size,
-)
+from tbxforms.layout import Size
 from tests.forms import FieldsetForm
 from tests.utils import render_form
 
@@ -19,18 +15,14 @@ def test_basic_layout(snapshot_html):
 def test_show_legend_as_heading(snapshot_html):
     """Verify the field legend can be displayed as the page heading."""
     form = FieldsetForm()
-    form.helper.layout = Layout(
-        Fieldset("name", "email", legend="Contact", legend_tag="h1")
-    )
+    form.helper[0].update_attributes(legend_tag="h1")
     assert render_form(form) == snapshot_html
 
 
 def test_change_legend_size(snapshot_html):
     """Verify size of the field legend can be changed from the default."""
     form = FieldsetForm()
-    form.helper.layout = Layout(
-        Fieldset("name", "email", legend="Contact", legend_size=Size.LARGE)
-    )
+    form.helper[0].update_attributes(legend_size=Size.LARGE)
     assert render_form(form) == snapshot_html
 
 
@@ -51,5 +43,5 @@ def test_css_id(snapshot_html):
 def test_attribute(snapshot_html):
     """Verify the extra attributes can be added."""
     form = FieldsetForm()
-    form.helper.layout = Layout(Fieldset(key="value"))
+    form.helper[0].update_attributes(key="value")
     assert render_form(form) == snapshot_html