+++ tests/layout/test_radios.py
@@ -36,34 +36,28 @@ def test_choices(snapshot_html):
 def test_small(snapshot_html):
     """Verify size of the radio buttons can be changed."""
     form = RadiosForm()
-    form.helper.layout = Layout(
-        Field("method", context={"radios_small": True})
-    )
+    form.helper.layout = Layout(Field.radios("method", small=True))
     assert render_form(form) == snapshot_html
 
 
 def test_inline(snapshot_html):
     """Verify radio buttons can be displayed in a row."""
     form = RadiosForm()
-    form.helper.layout = Layout(
-        Field("method", context={"radios_inline": True})
-    )
+    form.helper.layout = Layout(Field.radios("method", inline=True))
     assert render_form(form) == snapshot_html
 
 
 def test_show_legend_as_heading(snapshot_html):
     """Verify the field legend can be displayed as the page heading."""
     form = RadiosForm()
-    form.helper.layout = Layout(Field("method", context={"legend_tag": "h1"}))
+    form.helper.layout = Layout(Field.radios("method", legend_tag="h1"))
     assert render_form(form) == snapshot_html
 
 
 def test_change_legend_size(snapshot_html):
     """Verify size of the field legend can be changed from the default."""
     form = RadiosForm()
-    form.helper.layout = Layout(
-        Field("method", context={"legend_size": Size.for_legend("l")})
-    )
+    form.helper.layout = Layout(Field.radios("method", legend_size=Size.LARGE))
     assert render_form(form) == snapshot_html