+++ tests/layout/test_select.py
@@ -27,16 +27,14 @@ def test_validation_error_attributes(snapshot_html):
 def test_show_label_as_heading(snapshot_html):
     """Verify the field label can be displayed as the page heading."""
     form = SelectForm()
-    form.helper.layout = Layout(Field("method", context={"label_tag": "h1"}))
+    form.helper.layout = Layout(Field.select("method", label_tag="h1"))
     assert render_form(form) == snapshot_html
 
 
 def test_change_label_size(snapshot_html):
     """Verify size of the field label can be changed from the default."""
     form = SelectForm()
-    form.helper.layout = Layout(
-        Field("method", context={"label_size": Size.for_label("l")})
-    )
+    form.helper.layout = Layout(Field.select("method", label_size=Size.LARGE))
     assert render_form(form) == snapshot_html