+++ CHANGELOG.md
@@ -25,3 +25,7 @@
 
 -   Use snapshot testing plugin (syrupy) for component rendering tests instead of HTML fixtures
 -   form.helper (`FormHelper`) changed from a static `@property` to the form's `__init__` method to allow changes at runtime.
+
+### Fixed
+
+-   `Field.select` label size and tag can be changed.